+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Procedure when bumping the version number:
 
 ### Fixed
 - `!uptime` minute count
+- Instance reconnecting after encountering a 404 (it now stops and logs an error)
 
 ## v0.3.0 - 2023-02-11